Delete operations are modified and response status is added.

+++ src/main/java/com/everestengineering/employeeportalapplication/controllers/EmployeeController.java
@@ -15,33 +15,51 @@ import java.util.Optional;
 @RequestMapping("/api/employees")
 public class EmployeeController {
     private final EmployeeService employeeService;
+
     @GetMapping(value = "")
     public List<Employee> getAllEmployees() {
         return employeeService.getAllEmployees();
     }
+
     @GetMapping(value = "/{id}")
     public ResponseEntity<Optional<Employee>> getEmployeeById(@PathVariable(name = "id") Long id) {
         final Optional<Employee> Employee = employeeService.getEmployeeById(id);
-        if(Employee.isPresent()) {
+        if (Employee.isPresent()) {
             return ResponseEntity.ok(Employee);
         } else {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
     }
+
     @PostMapping(value = "")
     public ResponseEntity<Employee> createEmployee(@RequestBody Employee employee) {
         final Employee savedEmployee = employeeService.addEmployee(employee);
         return ResponseEntity.status(HttpStatus.CREATED).body(savedEmployee);
     }
+
     @PutMapping("/{id}")
     public Employee updateEmployee(@PathVariable(name = "id") Long id, @RequestBody Employee employee) {
-        return employeeService.updateEmployee(id,employee);
-    }
-    @DeleteMapping("/{id}")
-    public void deleteEmployee(@PathVariable(name = "id") Long id) {
-        employeeService.delete(id);
+        return employeeService.updateEmployee(id, employee);
     }
 
+    @DeleteMapping("")
+    public ResponseEntity<Void> deleteAllEmployees() {
+        if (employeeService.getAllEmployees().size() != 0) {
+            employeeService.deleteAllEmployees();
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteEmployee(@PathVariable(name = "id") Long id) {
+        if (employeeService.checkEmployeeId(id)) {
+            employeeService.delete(id);
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+
+    }
 
 
 }
